split file core.py to three files
split the file core to three files:
- core.py - containing only functions that can be used by users directly
- backtest.py - containing functions for backtesting
- optimization.py - containing functions for optimizaiton
files are not tested yet
