+++ qteasy/core.py
@@ -11,987 +11,26 @@
 import os
 import pandas as pd
 import numpy as np
-import time
-import math
 from warnings import warn
-from numba import njit
 
-from concurrent.futures import ProcessPoolExecutor, as_completed
 import datetime
 
 import qteasy
-from .history import get_history_panel, HistoryPanel, stack_dataframes
-from .utilfuncs import sec_to_duration, progress_bar, str_to_list, regulate_date_format, match_ts_code
+from .history import get_history_panel, HistoryPanel
+from .utilfuncs import str_to_list, regulate_date_format, match_ts_code
 from .utilfuncs import next_market_trade_day
 from .utilfuncs import AVAILABLE_ASSET_TYPES, _partial_lev_ratio
-from .space import Space, ResultPool
-from .finance import CashPlan, get_selling_result, get_purchase_result, set_cost
+from .finance import CashPlan
 from .qt_operator import Operator
 from .visual import _plot_loop_result, _print_loop_result, _print_test_result
 from .visual import _plot_test_result
-from .evaluate import evaluate, performance_statistics
 from ._arg_validators import _update_config_kwargs, ConfigDict
-
 from ._arg_validators import QT_CONFIG, _vkwargs_to_text
+from .optimization import _evaluate_all_parameters, _evaluate_one_parameter
 
 
 # TODO: for v1.1:
 #  reduce the size of this file, split it into several files
-
-@njit
-def _loop_step(signal_type: int,
-               own_cash: float,
-               own_amounts: np.ndarray,
-               available_cash: float,
-               available_amounts: np.ndarray,
-               op: np.ndarray,
-               prices: np.ndarray,
-               buy_fix: float,
-               sell_fix: float,
-               buy_rate: float,
-               sell_rate: float,
-               buy_min: float,
-               sell_min: float,
-               slipage: float,
-               pt_buy_threshold: float,
-               pt_sell_threshold: float,
-               maximize_cash_usage: bool,
-               long_pos_limit: float,
-               short_pos_limit: float,
-               allow_sell_short: bool,
-               moq_buy: float,
-               moq_sell: float) -> tuple:
-    """ 对同一批交易进行处理，采用向量化计算以提升效率
-        接受交易信号、交易价格以及期初可用现金和可用股票等输入，加上交易费率等信息计算交易后
-        的现金和股票变动值、并计算交易费用
-
-    Parameters
-    ----------
-    signal_type: int
-        信号类型:
-            0 - PT signal
-            1 - PS signal
-            2 - VS signal
-    own_cash: float
-        本次交易开始前持有的现金余额（包括交割中的现金）
-    own_amounts: np.ndarray
-        本次交易开始前持有的资产份额（包括交割中的资产份额）
-    available_cash: np.ndarray
-        本次交易开始前账户可用现金余额（交割中的现金不计入余额）
-    available_amounts: np.ndarray:
-        交易开始前各个股票的可用数量余额（交割中的股票不计入余额）
-    op: np.ndarray
-        本次交易的个股交易信号清单
-    prices: np.ndarray，
-        本次交易发生时各个股票的交易价格
-    buy_fix: float
-        交易成本：固定买入费用
-    sell_fix: float
-        交易成本：固定卖出费用
-    buy_rate: float
-        交易成本：固定买入费率
-    sell_rate: float
-        交易成本：固定卖出费率
-    buy_min: float
-        交易成本：最低买入费用
-    sell_min: float
-        交易成本：最低卖出费用
-    pt_buy_threshold: object Cost
-        当交易信号类型为PT时，用于计算买入/卖出信号的强度阈值
-    pt_sell_threshold: object Cost
-        当交易信号类型为PT时，用于计算买入/卖出信号的强度阈值
-    maximize_cash_usage: object Cost
-        True:   先卖后买模式
-        False:  先买后卖模式
-    long_pos_limit: float
-        允许建立的多头总仓位与净资产的比值，默认值1.0，表示最多允许建立100%多头仓位
-    short_pos_limit: float
-        允许建立的空头总仓位与净资产的比值，默认值-1.0，表示最多允许建立100%空头仓位
-    allow_sell_short: bool
-        True:   允许买空卖空
-        False:  默认值，只允许买入多头仓位
-    moq_buy: float:
-        投资产品最小买入交易单位，moq为0时允许交易任意数额的金融产品，moq不为零时允许交易的产品数量是moq的整数倍
-    moq_sell: float:
-        投资产品最小买入交易单位，moq为0时允许交易任意数额的金融产品，moq不为零时允许交易的产品数量是moq的整数倍
-
-    Returns
-    -------
-    tuple: (cash_gained, cash_spent, amounts_purchased, amounts_sold, fee)
-        cash_gained:        float, 本批次交易中获得的现金增加额
-        cash_spent:         float, 本批次交易中共花费的现金总额
-        amounts_purchased:  ndarray, 交易后每个个股账户中的股份增加数量
-        amounts_sold:       ndarray, 交易后每个个股账户中的股份减少数量
-        fee:                float, 本次交易总费用，包括卖出的费用和买入的费用
-    """
-
-    # 0, 当本期交易信号全为0，且信号类型为1或2时，退出计算，因为此时
-    # 的买卖行为仅受交易信号控制，交易信号全为零代表不交易，但是如果交
-    # 易信号为0时，代表持仓目标为0，此时有可能会有卖出交易，因此不能退
-    # 出计算
-    # if np.all(op == 0) and (signal_type > 0):
-    #     # 返回0代表获得和花费的现金，返回全0向量代表买入和卖出的股票
-    #     # 因为正好op全为0，因此返回op即可
-    #     return np.zeros_like(op), np.zeros_like(op), np.zeros_like(op), np.zeros_like(op), np.zeros_like(op)
-
-    # 1,计算期初资产总额：交易前现金及股票余额在当前价格下的资产总额
-    pre_values = own_amounts * prices
-    total_value = own_cash + pre_values.sum()
-
-    # 2,制定交易计划，生成计划买入金额和计划卖出数量
-    if signal_type == 0:
-        # signal_type 为PT，比较当前持仓与计划持仓的差额，再生成买卖数量
-        ptbt = pt_buy_threshold
-        ptst = -pt_sell_threshold
-        # 计算当前持仓与目标持仓之间的差额
-        pre_position = pre_values / total_value
-        position_diff = op - pre_position
-        # 当不允许买空卖空操作时，只需要考虑持有股票时卖出或买入，即开多仓和平多仓
-        # 当持有份额大于零时，平多仓：卖出数量 = 仓位差 * 持仓份额，此时持仓份额需大于零
-        amounts_to_sell = np.where((position_diff < ptst) & (own_amounts > 0),
-                                   position_diff / pre_position * own_amounts,
-                                   0.)
-        # 当持有份额不小于0时，开多仓：买入金额 = 仓位差 * 当前总资产，此时不能持有空头头寸
-        cash_to_spend = np.where((position_diff > ptbt) & (own_amounts >= 0),
-                                 position_diff * total_value,
-                                 0.)
-        # 当允许买空卖空时，允许开启空头头寸：
-        if allow_sell_short:
-
-            # 当持有份额小于等于零且交易信号为负，开空仓：买入空头金额 = 仓位差 * 当前总资产，此时持有份额为0
-            cash_to_spend += np.where((position_diff < ptst) & (own_amounts <= 0),
-                                      position_diff * total_value,
-                                      0.)
-            # 当持有份额小于0（即持有空头头寸）且交易信号为正时，平空仓：卖出空头数量 = 仓位差 * 当前持有空头份额
-            amounts_to_sell += np.where((position_diff > ptbt) & (own_amounts < 0),
-                                        position_diff / pre_position * own_amounts,
-                                        0.)
-
-    elif signal_type == 1:
-        # signal_type 为PS，根据目前的持仓比例和期初资产总额生成买卖数量
-        # 当不允许买空卖空操作时，只需要考虑持有股票时卖出或买入，即开多仓和平多仓
-        # 当持有份额大于零时，平多仓：卖出数量 =交易信号 * 持仓份额，此时持仓份额需大于零
-        amounts_to_sell = np.where((op < 0) & (own_amounts > 0), op * own_amounts, 0.)
-        # 当持有份额不小于0时，开多仓：买入金额 =交易信号 * 当前总资产，此时不能持有空头头寸
-        cash_to_spend = np.where((op > 0) & (own_amounts >= 0), op * total_value, 0.)
-
-        # 当允许买空卖空时，允许开启空头头寸：
-        if allow_sell_short:
-
-            # 当持有份额小于等于零且交易信号为负，开空仓：买入空头金额 = 交易信号 * 当前总资产
-            cash_to_spend += np.where((op < 0) & (own_amounts <= 0), op * total_value, 0.)
-            # 当持有份额小于0（即持有空头头寸）且交易信号为正时，平空仓：卖出空头数量 = 交易信号 * 当前持有空头份额
-            amounts_to_sell -= np.where((op > 0) & (own_amounts < 0), op * own_amounts, 0.)
-
-    elif signal_type == 2:
-        # signal_type 为VS，交易信号就是计划交易的股票数量，符号代表交易方向
-        # 当不允许买空卖空操作时，只需要考虑持有股票时卖出或买入，即开多仓和平多仓
-        # 当持有份额大于零时，卖出多仓：卖出数量 = 信号数量，此时持仓份额需大于零
-        amounts_to_sell = np.where((op < 0) & (own_amounts > 0), op, 0.)
-        # 当持有份额不小于0时，买入多仓：买入金额 = 信号数量 * 资产价格，此时不能持有空头头寸，必须为空仓或多仓
-        cash_to_spend = np.where((op > 0) & (own_amounts >= 0), op * prices, 0.)
-
-        # 当允许买空卖空时，允许开启空头头寸：
-        if allow_sell_short:
-            # 当持有份额小于等于零且交易信号为负，买入空仓：买入空头金额 = 信号数量 * 资产价格
-            cash_to_spend += np.where((op < 0) & (own_amounts <= 0), op * prices, 0.)
-            # 当持有份额小于0（即持有空头头寸）且交易信号为正时，卖出空仓：卖出空头数量 = 交易信号 * 当前持有空头份额
-            amounts_to_sell -= np.where((op > 0) & (own_amounts < 0), -op, 0.)
-
-    else:
-        raise ValueError('Invalid signal_type')
-
-    # 3, 批量提交股份卖出计划，计算实际卖出份额与交易费用。
-
-    # 如果不允许卖空交易，则需要更新股票卖出计划数量
-    if not allow_sell_short:
-        amounts_to_sell = - np.fmin(-amounts_to_sell, available_amounts)
-    amount_sold, cash_gained, fee_selling = get_selling_result(
-            prices=prices,
-            a_to_sell=amounts_to_sell,
-            moq=moq_sell,
-            buy_fix=buy_fix,
-            sell_fix=sell_fix,
-            buy_rate=buy_rate,
-            sell_rate=sell_rate,
-            buy_min=buy_min,
-            sell_min=sell_min,
-            slipage=slipage
-    )
-
-    if maximize_cash_usage:
-        # 仅当现金交割期为0，且希望最大化利用同批交易产生的现金时，才调整现金余额
-        # 现金余额 = 期初现金余额 + 本次出售资产获得现金总额
-        available_cash += cash_gained.sum()
-
-    # 调整处理cash_to_spend
-    # 初步估算按照交易清单买入资产所需要的现金，如果超过持有现金，则按比例降低买入金额
-    abs_cash_to_spend = np.abs(cash_to_spend)
-
-    if np.all(abs_cash_to_spend < 0.01):
-        # 如果所有买入计划绝对值都小于1分钱，则直接跳过后续的计算
-        return cash_gained, np.zeros_like(op), np.zeros_like(op), amount_sold, fee_selling
-
-    # 分别处理买入金额中的多头买入和空头买入部分，分别计算当前持有的多头和空头仓位
-    pos_cash_to_spend = np.where(cash_to_spend > 0.01, cash_to_spend, 0)
-    total_pos_cash_to_spend = pos_cash_to_spend.sum()
-    neg_cash_to_spend = np.where(cash_to_spend < -0.01, cash_to_spend, 0)
-    total_neg_cash_to_spend = neg_cash_to_spend.sum()
-    next_own_amounts = own_amounts + amount_sold
-
-    # 计算允许用于买入多头份额的最大金额
-    current_long_pos = np.where(next_own_amounts > 0, next_own_amounts * prices, 0).sum() / total_value
-    max_pos_cash_to_spend = (long_pos_limit - current_long_pos) * total_value
-
-    if long_pos_limit <= 1.0:
-        max_pos_cash_to_spend = min(max_pos_cash_to_spend, available_cash)
-
-    if total_pos_cash_to_spend > max_pos_cash_to_spend:
-        # 如果计划买入多头现金超过允许买入最大金额，按比例降低分配给每个拟买入多头资产的现金
-        pos_cash_to_spend = pos_cash_to_spend / total_pos_cash_to_spend * max_pos_cash_to_spend
-
-    if allow_sell_short:
-        # 只有当allow_sell_short的时候才去考察允许持有的空头仓位限制
-        current_short_pos = np.where(next_own_amounts < 0, next_own_amounts * prices, 0).sum() / total_value
-        max_neg_cash_to_spend = (short_pos_limit - current_short_pos) * total_value
-
-        if total_neg_cash_to_spend < max_neg_cash_to_spend:
-            # 如果计划买入空头现金超过允许买入最大金额，按比例调降拟买入空头资产的现金
-            neg_cash_to_spend = neg_cash_to_spend / total_neg_cash_to_spend * max_neg_cash_to_spend
-
-    cash_to_spend = pos_cash_to_spend + neg_cash_to_spend
-
-    # 批量提交股份买入计划，计算实际买入的股票份额和交易费用
-    # 由于已经提前确认过现金总额，因此不存在买入总金额超过持有现金的情况
-    amount_purchased, cash_spent, fee_buying = get_purchase_result(
-            prices=prices,
-            cash_to_spend=cash_to_spend,
-            moq=moq_buy,
-            buy_fix=buy_fix,
-            sell_fix=sell_fix,
-            buy_rate=buy_rate,
-            sell_rate=sell_rate,
-            buy_min=buy_min,
-            sell_min=sell_min,
-            slipage=slipage
-    )
-
-    # 4, 计算购入资产产生的交易成本，买入资产和卖出资产的交易成本率可以不同，且每次交易动态计算
-    fee = fee_buying + fee_selling
-
-    return cash_gained, cash_spent, amount_purchased, amount_sold, fee
-
-
-def _get_complete_hist(looped_value: pd.DataFrame,
-                       h_list: HistoryPanel,
-                       benchmark_list: pd.DataFrame,
-                       with_price: bool = False) -> pd.DataFrame:
-    """完成历史交易回测后，填充完整的历史资产总价值清单，
-        同时在回测清单中填入参考价格数据，参考价格数据用于数据可视化对比，参考数据的来源为Config.benchmark_asset
-
-    Parameters
-    ----------
-    looped_value: pd.DataFrame
-        完成历史交易回测后生成的历史资产总价值清单，只有在操作日才有记录，非操作日没有记录
-    h_list: pd.DataFrame
-        完整的投资产品价格清单，包含所有投资产品在回测区间内每个交易日的价格
-    benchmark_list: pd.DataFrame
-        参考资产的历史价格清单，参考资产用于收益率的可视化对比，同时用于计算alpha、sharp等指标
-    with_price: boolean, default False
-        True时在返回的清单中包含历史价格，否则仅返回资产总价值
-
-    Returns
-    -------
-    looped_value: pd.DataFrame:
-    重新填充的完整历史交易日资产总价值清单，包含以下列：
-    - [share-x]:        多列，每种投资产品的持有份额数量
-    - cash:             期末现金金额
-    - fee:              当期交易费用（交易成本）
-    - value:            当期资产总额（现金总额 + 所有在手投资产品的价值总额）
-    """
-    # 获取价格清单中的投资产品列表
-    shares = h_list.shares  # 获取资产清单
-    try:
-        start_date = looped_value.index[0]  # 开始日期
-    except:
-        raise IndexError('index 0 is out of bounds for axis 0 with size 0')
-    looped_history = h_list.segment(start_date)  # 回测历史数据区间 = [开始日期:]
-    # 使用价格清单的索引值对资产总价值清单进行重新索引，重新索引时向前填充每日持仓额、现金额，使得新的
-    # 价值清单中新增的记录中的持仓额和现金额与最近的一个操作日保持一致，并消除nan值
-    hdates = looped_history.hdates
-    purchased_shares = looped_value[shares].reindex(hdates, method='ffill').fillna(0)
-    cashes = looped_value['cash'].reindex(hdates, method='ffill').fillna(0)
-    fees = looped_value['fee'].reindex(hdates).fillna(0)
-    looped_value = looped_value.reindex(hdates)
-    # 这里采用了一种看上去貌似比较奇怪的处理方式：
-    # 先为cashes、purchased_shares两个变量赋值，
-    # 然后再将上述两个变量赋值给looped_values的列中
-    # 这样看上去好像多此一举，为什么不能直接赋值，然而
-    # 经过测试，如果直接用下面的代码直接赋值，将无法起到
-    # 填充值的作用：
-    # looped_value.cash = looped_value.cash.reindex(dates, method='ffill')
-    looped_value[shares] = purchased_shares
-    looped_value.cash = cashes
-    looped_value.fee = looped_value['fee'].reindex(hdates).fillna(0)
-    looped_value['reference'] = benchmark_list.reindex(hdates).fillna(0)
-    # 重新计算整个清单中的资产总价值，生成pandas.Series对象，如果looped_history历史价格中包含多种价格，使用最后一种
-    decisive_prices = looped_history[-1].squeeze(axis=2).T
-    looped_value['value'] = (decisive_prices * looped_value[shares]).sum(axis=1) + looped_value['cash']
-    if with_price:  # 如果需要同时返回价格，则生成pandas.DataFrame对象，包含所有历史价格
-        share_price_column_names = [name + '_p' for name in shares]
-        looped_value[share_price_column_names] = looped_history[shares]
-    return looped_value
-
-
-def _merge_invest_dates(op_list: pd.DataFrame, invest: CashPlan) -> pd.DataFrame:
-    """将完成的交易信号清单与现金投资计划合并：
-
-    检查现金投资计划中的日期是否都存在于交易信号清单op_list中，如果op_list中没有相应日期时，当交易信号清单中没有相应日期时，添加空交易
-    信号，以便使op_list中存在相应的日期。
-
-    注意，在目前的设计中，要求invest中的所有投资日期都为交易日（意思是说，投资日期当天资产价格不为np.nan）
-    否则，将会导致回测失败（回测当天取到np.nan作为价格，引起总资产计算失败，引起连锁反应所有的输出结果均为np.nan。
-
-    需要在CashPlan投资计划类中增加合法性判断
-
-    Parameters
-    ----------
-    op_list: pd.DataFrame
-        交易信号清单，一个DataFrame。index为Timestamp类型
-    invest: qt.CashPlan
-        CashPlan对象，投资日期和投资金额
-    Returns
-    -------
-    op_list: pd.DataFrame
-        合并后的交易信号清单
-    """
-    if not isinstance(op_list, pd.DataFrame):
-        raise TypeError(f'operation list should be a pandas DataFrame, got {type(op_list)} instead')
-    if not isinstance(invest, CashPlan):
-        raise TypeError(f'invest plan should be a qteasy CashPlan, got {type(invest)} instead')
-    for date in invest.dates:
-        try:
-            op_list.loc[date]
-        except Exception:
-            op_list.loc[date] = 0
-    op_list.sort_index(inplace=True)
-    return op_list
-
-
-# TODO: 删除operator作为传入参数，仅处理回测结果，将回测结果传出
-#  函数后再处理为pandas.DataFrame，并在函数以外进行进一步的记录和处理，这里仅仅使用与回测相关
-#  的参数
-def apply_loop(operator: Operator,
-               trade_price_list: HistoryPanel,
-               start_idx: int = 0,
-               end_idx: int = None,
-               cash_plan: CashPlan = None,
-               cost_rate: dict = None,
-               moq_buy: float = 100.,
-               moq_sell: float = 1.,
-               inflation_rate: float = 0.03,
-               pt_signal_timing: str = 'lazy',
-               pt_buy_threshold: float = 0.1,
-               pt_sell_threshold: float = 0.1,
-               cash_delivery_period: int = 0,
-               stock_delivery_period: int = 0,
-               allow_sell_short: bool = False,
-               long_pos_limit: float = 1.0,
-               short_pos_limit: float = -1.0,
-               max_cash_usage: bool = False,
-               trade_log: bool = False,
-               price_priority_list: list = None) -> tuple:
-    """使用Numpy快速迭代器完成整个交易清单在历史数据表上的模拟交易，并输出每次交易后持仓、
-        现金额及费用，输出的结果可选
-
-    Parameters
-    ----------
-    operator: Operator
-        用于生成交易信号(realtime模式)，预先生成的交易信号清单或清单相关信息也从中读取
-    start_idx: int, Default: 0
-        模拟交易从交易清单的该序号开始循环
-    end_idx: int, Default: None
-        模拟交易到交易清单的该序号为止
-    trade_price_list: object HistoryPanel
-        完整历史价格清单，数据的频率由freq参数决定
-    cash_plan: CashPlan: Default: None
-        资金投资计划，CashPlan对象
-    cost_rate: dict: Default: None
-        交易成本率，包含交易费、滑点及冲击成本
-    moq_buy: float：Default: 100
-        每次交易买进的最小份额单位
-    moq_sell: float: Default: 1
-        每次交易卖出的最小份额单位
-    inflation_rate: float, Default: 0.03
-        现金的时间价值率，如果>0，则现金的价值随时间增长，增长率为inflation_rate
-    pt_signal_timing: str, {'lazy', 'eager', 'aggressive'}  # TODO: 增加参数值 'aggressive' 的alias 'eager'
-        控制PT模式下交易信号产生的时机
-    pt_buy_threshold: float, Default: 0.1
-        PT买入信号阈值，只有当实际持仓与目标持仓的差值大于该阈值时，才会产生买入信号
-    pt_sell_threshold: flaot, Default: 0.1
-        PT卖出信号阈值，只有当实际持仓与目标持仓的差值小于该阈值时，才会产生卖出信号
-    cash_delivery_period: int, Default: 0
-        现金交割周期，默认值为0，单位为天。
-    stock_delivery_period: int, Default: 0
-        股票交割周期，默认值为0，单位为天。
-    allow_sell_short: bool, Default: False
-        是否允许卖空操作，如果不允许卖空，则卖出的数量不能超过持仓数量
-    long_pos_limit: float, Default: 1.0
-        允许持有的最大多头仓位比例
-    short_pos_limit: flaot, Default: -1.0
-        允许持有的最大空头仓位比例
-    max_cash_usage: bool, Default: False
-        是否最大化利用现金，如果为True，则在每次交易时，会将卖出股票的现金用于买入股票
-    trade_log: bool: Default: False
-        为True时，输出回测详细日志为csv格式的表格
-    price_priority_list: list, Default: None
-        各交易价格的执行顺序列表，列表中0～N数字代表operator中的各个交易价格的序号，各个交易价格
-        将按照列表中的序号顺序执行
-
-    Returns
-    -------
-    tuple: (loop_results, op_log_matrix, op_summary_matrix, op_list_bt_indices)
-    - loop_results:        用于生成交易结果的数据，如持仓数量、交易费用、持有现金以及总资产
-    - op_log_matrix:       用于生成详细交易记录的数据，包含每次交易的详细交易信息，如持仓、成交量、成交价格、现金变动、交易费用等等
-    - op_summary_matrix:   用于生成详细交易记录的补充数据，包括投入资金量、资金变化量等
-    - op_list_bt_indices:  交易清单中实际参加回测的行序号
-    """
-    if moq_buy == 0:
-        assert moq_sell == 0, f'ValueError, if "trade_batch_size" is 0, then ' \
-                              f'"sell_batch_size" should also be 0'
-    if (moq_buy != 0) and (moq_sell != 0):
-        assert moq_buy % moq_sell == 0, \
-            f'ValueError, the sell moq should be divisible by moq_buy, or there will be mistake'
-    signal_type = operator.signal_type_id
-    op_type = operator.op_type
-
-    # 获取交易信号的总行数、股票数量以及价格种类数量
-    # 在这里，交易信号的价格种类数量与交易价格的价格种类数量必须一致，且顺序也必须一致
-    op_list = None
-    if operator.op_type == 'batch':
-
-        op_list = operator.op_list
-
-    looped_dates = operator.op_list_hdates
-    share_count, op_count, price_type_count = operator.op_list_shape
-    if end_idx is None:
-        end_idx = op_count
-    # 为防止回测价格数据中存在Nan值，需要首先将Nan值替换成0，否则将造成错误值并一直传递到回测历史最后一天
-    price = trade_price_list.ffill(0).values
-    # 获取每一个资金投入日在历史时间序列中的位置
-    investment_date_pos = np.searchsorted(looped_dates, cash_plan.dates)
-    invest_dict = cash_plan.to_dict(investment_date_pos)
-    # 解析交易费率参数：
-    buy_fix = cost_rate['buy_fix']
-    sell_fix = cost_rate['sell_fix']
-    buy_rate = cost_rate['buy_rate']
-    sell_rate = cost_rate['sell_rate']
-    buy_min = cost_rate['buy_min']
-    sell_min = cost_rate['sell_min']
-    slipage = cost_rate['slipage']
-    # 确定是否属于PT+lazy的情形
-    pt_and_lazy = (signal_type == 0) and (pt_signal_timing == 'lazy')
-
-    # 检查信号清单，生成清单回测序号，生成需要跳过的交易信号的列表。这个列表有2维，分别代表每日/每回测价格信号是否可以跳过
-    if operator.op_type == 'stepwise':
-        # 在stepwise模式下，每一天都需要回，回测所有交易信号行
-        skip_op_signal = np.zeros(shape=(op_count, price_type_count), dtype='bool')
-    elif signal_type in [1, 2]:
-        # 否则，在batch模式下，PS/VS模式下跳过没有信号的交易日, 并确保第一个回测日不为True
-        skip_op_signal = np.all(op_list == 0, axis=0)
-        skip_op_signal[start_idx, :] = False
-    elif pt_and_lazy:
-        # 或者，在batch模式下，PT模式下跳过信号没有发生变化的交易日，但确保第一个回测日不为True
-        signal_diff = op_list - np.roll(op_list, 1, axis=1)
-        skip_op_signal = np.all(signal_diff == 0, axis=0)
-        skip_op_signal[start_idx, :] = False
-    else:
-        # 否则，在batch/PT/aggressive模式下，回测所有交易信号行
-        skip_op_signal = np.zeros(shape=(op_count, price_type_count), dtype='bool')
-    # 确定bt回测计算的范围
-    op_list_bt_indices = np.array(range(start_idx, end_idx))
-    # 如果inflation_rate > 0 则还需要计算所有有交易信号的日期相对前一个交易信号日的现金增长比率，这个比率与两个交易信号日之间的时间差有关
-    inflation_factors = np.ones_like(op_list_bt_indices, dtype='float')
-    additional_invest = 0.
-    # 如果inflation_rate不为0，则需要计算每一个bt回测交易日相对上一日的现金增值幅度，用于计算现金增值
-    if inflation_rate > 0:
-        # TODO: 考虑把下面的计算numba化
-        # 在不同的datafreq下，相差一个idx不一定代表相差一天，因此需要计算每个idx之间实际相差的天数
-        bt_index_days = pd.to_datetime(looped_dates)[op_list_bt_indices]
-        days_difference = np.array((bt_index_days - np.roll(bt_index_days, 1)).days)
-        daily_ir = 1 + inflation_rate / 365.  # 由于这几计算的是两个日期之间的自然天数之差，因此日利率需要用ir/365计算
-        inflation_factors = daily_ir ** days_difference
-        inflation_factors[0] = 1.  # 使用np.roll计算后的第一个值是错误值，需要修正为1
-
-    # 决定交易中是否最大化使用现金
-    maximize_cash_usage = max_cash_usage and (cash_delivery_period == 0)
-    # 保存trade_log_table数据：
-    op_log_add_invest = []
-    op_log_cash = []
-    op_log_available_cash = []
-    op_log_value = []
-    op_log_matrix = []
-    prev_date = 0
-
-    if (op_type == 'batch') and (not trade_log):
-        # batch模式下调用apply_loop_core函数:
-        looped_day_indices = list(pd.to_datetime(pd.to_datetime(looped_dates).date).astype('int'))
-        # 2, 将invset_dict处理为两个列表：invest_date_indices, invest_amount，因为invest_dict无法被Numba处理
-        investment_date_pos = list(investment_date_pos)
-        invest_amounts = list(invest_dict.values())
-        cashes, fees, values, amounts_matrix = apply_loop_core(share_count,
-                                                               looped_day_indices,
-                                                               inflation_factors,
-                                                               investment_date_pos,
-                                                               invest_amounts,
-                                                               price,
-                                                               op_list,
-                                                               signal_type,
-                                                               op_list_bt_indices,
-                                                               skip_op_signal,
-                                                               buy_fix,
-                                                               sell_fix,
-                                                               buy_rate,
-                                                               sell_rate,
-                                                               buy_min,
-                                                               sell_min,
-                                                               slipage,
-                                                               moq_buy,
-                                                               moq_sell,
-                                                               inflation_rate,
-                                                               pt_buy_threshold,
-                                                               pt_sell_threshold,
-                                                               cash_delivery_period,
-                                                               stock_delivery_period,
-                                                               allow_sell_short,
-                                                               long_pos_limit,
-                                                               short_pos_limit,
-                                                               maximize_cash_usage,
-                                                               price_priority_list)
-    else:
-        # 初始化计算结果列表
-        own_cash = 0.  # 持有现金总额，期初现金总额总是0，在回测过程中到现金投入日时再加入现金
-        available_cash = 0.  # 每期可用现金总额
-        own_amounts = np.zeros(shape=(share_count,))  # 投资组合中各个资产的持有数量，初始值为全0向量
-        available_amounts = np.zeros(shape=(share_count,))  # 每期可用的资产数量
-        cash_delivery_queue = []  # 用于模拟现金交割延迟期的定长队列
-        stock_delivery_queue = []  # 用于模拟股票交割延迟期的定长队列
-        cashes = []  # 中间变量用于记录各个资产买入卖出时消耗或获得的现金
-        fees = []  # 交易费用，记录每个操作时点产生的交易费用
-        values = []  # 资产总价值，记录每个操作时点的资产和现金价值总和
-        amounts_matrix = []
-        total_value = 0
-        trade_data = np.zeros(shape=(share_count, 5))  # 交易汇总数据表，包含最近成交、交易价格、持仓数量、
-        # 持有现金等数据的数组，用于stepwise信号生成
-        recent_amounts_change = np.zeros(shape=(share_count,))  # 中间变量，保存最近的一次交易数量
-        recent_trade_prices = np.zeros(shape=(share_count,))  # 中间变量，保存最近一次的成交价格
-        result_count = 0  # 进行循环的次数
-        # 在stepwise模式下pt_and_lazy情形需要跳过某些交易信号，需要用到prev_op
-        prev_op = np.empty(shape=(share_count, price_type_count), dtype='float')
-        prev_op[:, :] = np.nan
-        for i in op_list_bt_indices:
-            # 对每一回合历史交易信号开始回测，每一回合包含若干交易价格上所有股票的交易信号
-            current_date = looped_dates[i].date()
-            sub_total_fee = 0
-            if inflation_rate > 0:
-                # 现金的价值随时间增长，需要依次乘以inflation 因子，且只有持有现金增值，新增的现金不增值
-                current_inflation_factor = inflation_factors[result_count]
-                own_cash *= current_inflation_factor
-                available_cash *= current_inflation_factor
-            if i in investment_date_pos:
-                # 如果在交易当天有资金投入，则将投入的资金加入可用资金池中
-                additional_invest = invest_dict[i]
-                own_cash += additional_invest
-                available_cash += additional_invest
-            for j in price_priority_list:
-                # 交易前将交割队列中达到交割期的现金完成交割
-                if ((prev_date != current_date) and
-                    (len(cash_delivery_queue) == cash_delivery_period)) or \
-                        (cash_delivery_period == 0):
-                    if len(cash_delivery_queue) > 0:
-                        cash_delivered = cash_delivery_queue.pop(0)
-                        available_cash += cash_delivered
-                # 交易前将交割队列中达到交割期的资产完成交割
-                if ((prev_date != current_date) and
-                    (len(stock_delivery_queue) == stock_delivery_period)) or \
-                        (stock_delivery_period == 0):
-                    if len(stock_delivery_queue) > 0:
-                        stock_delivered = stock_delivery_queue.pop(0)
-                        available_amounts += stock_delivered
-                # 调用loop_step()函数，计算本轮交易的现金和股票变动值以及总交易费用
-                current_prices = price[:, result_count, j]
-                if op_type == 'stepwise':
-                    # 在realtime模式下，准备trade_data并计算下一步的交易信号
-                    trade_data[:, 0] = own_amounts
-                    trade_data[:, 1] = available_amounts
-                    trade_data[:, 2] = current_prices
-                    trade_data[:, 3] = recent_amounts_change
-                    trade_data[:, 4] = recent_trade_prices
-                    current_op = operator.create_signal(
-                            trade_data=trade_data,
-                            sample_idx=i,
-                            price_type_idx=j
-                    )
-                    if pt_and_lazy and np.all(prev_op[:, j] == current_op):
-                        skip_op_signal[i, j] = True
-                    prev_op[:, j] = current_op
-
-                elif op_type == 'batch':
-                    # 在batch模式下，直接从批量生成的交易信号清单中读取下一步交易信号
-                    current_op = op_list[:, i, j]
-                else:
-                    # 其他不合法的op_type
-                    raise TypeError(f'invalid op_type!')
-                # 处理需要回测的交易信号，只有需要回测的信号才送入loop_step，否则直接生成五组全0结果
-                if skip_op_signal[i, j]:
-                    cash_gained = np.zeros_like(current_op)
-                    cash_spent = np.zeros_like(current_op)
-                    amount_purchased = np.zeros_like(current_op)
-                    amount_sold = np.zeros_like(current_op)
-                    fee = np.zeros_like(current_op)
-                else:
-                    cash_gained, cash_spent, amount_purchased, amount_sold, fee = _loop_step(
-                            signal_type=signal_type,
-                            own_cash=own_cash,
-                            own_amounts=own_amounts,
-                            available_cash=available_cash,
-                            available_amounts=available_amounts,
-                            op=current_op,
-                            prices=current_prices,
-                            buy_fix=buy_fix,
-                            sell_fix=sell_fix,
-                            buy_rate=buy_rate,
-                            sell_rate=sell_rate,
-                            buy_min=buy_min,
-                            sell_min=sell_min,
-                            slipage=slipage,
-                            pt_buy_threshold=pt_buy_threshold,
-                            pt_sell_threshold=pt_sell_threshold,
-                            maximize_cash_usage=maximize_cash_usage,
-                            allow_sell_short=allow_sell_short,
-                            long_pos_limit=long_pos_limit,
-                            short_pos_limit=short_pos_limit,
-                            moq_buy=moq_buy,
-                            moq_sell=moq_sell
-                    )
-                # 获得的现金进入交割队列，根据日期的变化确定是新增现金交割还是累加现金交割
-                if (prev_date != current_date) or (cash_delivery_period == 0):
-                    cash_delivery_queue.append(cash_gained.sum())
-                else:
-                    cash_delivery_queue[-1] += cash_gained.sum()
-
-                # 获得的资产进入交割队列，根据日期的变化确定是新增资产交割还是累加资产交割
-                if (prev_date != current_date) or (stock_delivery_period == 0):
-                    stock_delivery_queue.append(amount_purchased)
-                else:  # if prev_date == current_date
-                    stock_delivery_queue[-1] += amount_purchased
-
-                prev_date = current_date
-                # 持有现金、持有股票用于计算本期的总价值
-                available_cash += cash_spent.sum()
-                available_amounts += amount_sold
-                cash_changed = cash_gained + cash_spent
-                own_cash = own_cash + cash_changed.sum()
-                amount_changed = amount_sold + amount_purchased
-                own_amounts = own_amounts + amount_changed
-                total_stock_values = (own_amounts * current_prices)
-                total_stock_value = total_stock_values.sum()
-                total_value = total_stock_value + own_cash
-                sub_total_fee += fee.sum()
-                # 生成trade_log所需的数据，采用串列式表格排列：
-                if trade_log:
-                    rnd = np.round
-                    op_log_matrix.append(rnd(current_op, 3))
-                    op_log_matrix.append(rnd(current_prices, 3))
-                    op_log_matrix.append(rnd(amount_changed, 3))
-                    op_log_matrix.append(rnd(cash_changed, 3))
-                    op_log_matrix.append(rnd(fee, 3))
-                    op_log_matrix.append(rnd(own_amounts, 3))
-                    op_log_matrix.append(rnd(available_amounts, 3))
-                    op_log_matrix.append(rnd(total_stock_values, 3))
-                    op_log_add_invest.append(rnd(additional_invest, 3))
-                    additional_invest = 0.
-                    op_log_cash.append(rnd(own_cash, 3))
-                    op_log_available_cash.append(rnd(available_cash, 3))
-                    op_log_value.append(rnd(total_value, 3))
-
-            # 保存计算结果
-            cashes.append(own_cash)
-            fees.append(sub_total_fee)
-            values.append(total_value)
-            amounts_matrix.append(own_amounts)
-            result_count += 1
-
-    loop_results = (amounts_matrix, cashes, fees, values)
-    op_summary_matrix = (op_log_add_invest, op_log_cash, op_log_available_cash, op_log_value)
-    return loop_results, op_log_matrix, op_summary_matrix, op_list_bt_indices
-
-
-@njit
-def apply_loop_core(share_count,
-                    looped_dates,
-                    inflation_factors,
-                    investment_date_pos,
-                    invest_amounts,
-                    price,
-                    op_list,
-                    signal_type,
-                    op_list_bt_indices,
-                    skip_op_signal,
-                    buy_fix: float,
-                    sell_fix: float,
-                    buy_rate: float,
-                    sell_rate: float,
-                    buy_min: float,
-                    sell_min: float,
-                    slipage: float,
-                    moq_buy: float,
-                    moq_sell: float,
-                    inflation_rate: float,
-                    pt_buy_threshold: float,
-                    pt_sell_threshold: float,
-                    cash_delivery_period: int,
-                    stock_delivery_period: int,
-                    allow_sell_short: bool,
-                    long_pos_limit: float,
-                    short_pos_limit: float,
-                    max_cash_usage: bool,
-                    price_priority_list: list):
-    """ apply_loop的核心function,不含任何numba不支持的元素，仅包含batch模式下，
-        不需要生成trade_log的情形下运行核心循环的核心代码。
-        在符合要求的情况下，这部分代码以njit方式加速运行，实现提速
-
-    Returns
-    -------
-    """
-
-    # 初始化计算结果列表
-    own_cash = 0.  # 持有现金总额，期初现金总额总是0，在回测过程中到现金投入日时再加入现金
-    available_cash = 0.  # 每期可用现金总额
-    own_amounts = np.zeros(shape=(share_count,))  # 投资组合中各个资产的持有数量，初始值为全0向量
-    available_amounts = np.zeros(shape=(share_count,))  # 每期可用的资产数量
-    cash_delivery_queue = []  # 用于模拟现金交割延迟期的定长队列
-    stock_delivery_queue = []  # 用于模拟股票交割延迟期的定长队列
-    signal_count = len(op_list_bt_indices)
-    cashes = np.empty(shape=(signal_count,))  # 中间变量用于记录各个资产买入卖出时消耗或获得的现金
-    fees = np.empty(shape=(signal_count,))  # 交易费用，记录每个操作时点产生的交易费用
-    values = np.empty(shape=(signal_count,))  # 资产总价值，记录每个操作时点的资产和现金价值总和
-    amounts_matrix = np.empty(shape=(signal_count, share_count))
-    total_value = 0
-    prev_date = 0
-    investment_count = 0  # 用于正确读取每笔投资金额的计数器
-    result_count = 0  # 用于确保正确输出每笔交易结果的计数器
-
-    for i in op_list_bt_indices:
-        # 对每一回合历史交易信号开始回测，每一回合包含若干交易价格上所有股票的交易信号
-        current_date = looped_dates[i]
-        sub_total_fee = 0
-        if inflation_rate > 0:
-            # 现金的价值随时间增长，需要依次乘以inflation 因子，且只有持有现金增值，新增的现金不增值
-            own_cash *= inflation_factors[result_count]
-            available_cash *= inflation_factors[result_count]
-        if i in investment_date_pos:
-            # 如果在交易当天有资金投入，则将投入的资金加入可用资金池中
-            additional_invest = invest_amounts[investment_count]
-            own_cash += additional_invest
-            available_cash += additional_invest
-            investment_count += 1
-        for j in price_priority_list:
-            # 交易前将交割队列中达到交割期的现金完成交割
-            if ((prev_date != current_date) and
-                (len(cash_delivery_queue) == cash_delivery_period)) or \
-                    (cash_delivery_period == 0):
-                if len(cash_delivery_queue) > 0:
-                    cash_delivered = cash_delivery_queue.pop(0)
-                    available_cash += cash_delivered
-            # 交易前将交割队列中达到交割期的资产完成交割
-            if ((prev_date != current_date) and
-                (len(stock_delivery_queue) == stock_delivery_period)) or \
-                    (stock_delivery_period == 0):
-                if len(stock_delivery_queue) > 0:
-                    stock_delivered = stock_delivery_queue.pop(0)
-                    available_amounts += stock_delivered
-            # 调用loop_step()函数，计算本轮交易的现金和股票变动值以及总交易费用
-            current_prices = price[:, result_count, j]
-            current_op = op_list[:, i, j]
-            if skip_op_signal[i, j]:
-                cash_gained = np.zeros_like(current_op)
-                cash_spent = np.zeros_like(current_op)
-                amount_purchased = np.zeros_like(current_op)
-                amount_sold = np.zeros_like(current_op)
-                fee = np.zeros_like(current_op)
-            else:
-                cash_gained, cash_spent, amount_purchased, amount_sold, fee = _loop_step(
-                        signal_type=signal_type,
-                        own_cash=own_cash,
-                        own_amounts=own_amounts,
-                        available_cash=available_cash,
-                        available_amounts=available_amounts,
-                        op=current_op,
-                        prices=current_prices,
-                        buy_fix=buy_fix,
-                        sell_fix=sell_fix,
-                        buy_rate=buy_rate,
-                        sell_rate=sell_rate,
-                        buy_min=buy_min,
-                        sell_min=sell_min,
-                        slipage=slipage,
-                        pt_buy_threshold=pt_buy_threshold,
-                        pt_sell_threshold=pt_sell_threshold,
-                        maximize_cash_usage=max_cash_usage and cash_delivery_period == 0,
-                        allow_sell_short=allow_sell_short,
-                        long_pos_limit=long_pos_limit,
-                        short_pos_limit=short_pos_limit,
-                        moq_buy=moq_buy,
-                        moq_sell=moq_sell
-                )
-            # 获得的现金进入交割队列，根据日期的变化确定是新增现金交割还是累加现金交割
-            if (prev_date != current_date) or (cash_delivery_period == 0):
-                cash_delivery_queue.append(cash_gained.sum())
-            else:
-                cash_delivery_queue[-1] += cash_gained.sum()
-
-            # 获得的资产进入交割队列，根据日期的变化确定是新增资产交割还是累加资产交割
-            if (prev_date != current_date) or (stock_delivery_period == 0):
-                stock_delivery_queue.append(amount_purchased)
-            else:  # if prev_date == current_date
-                stock_delivery_queue[-1] += amount_purchased
-
-            prev_date = current_date
-            # 持有现金、持有股票用于计算本期的总价值
-            available_cash += cash_spent.sum()
-            available_amounts += amount_sold
-            cash_changed = cash_gained + cash_spent
-            own_cash = own_cash + cash_changed.sum()
-            amount_changed = amount_sold + amount_purchased
-            own_amounts = own_amounts + amount_changed
-            total_stock_values = (own_amounts * current_prices)
-            total_stock_value = total_stock_values.sum()
-            total_value = total_stock_value + own_cash
-            sub_total_fee += fee.sum()
-
-        # 保存计算结果
-        cashes[result_count] = own_cash
-        fees[result_count] = sub_total_fee
-        values[result_count] = total_value
-        amounts_matrix[result_count, :] = own_amounts
-        result_count += 1
-
-    return cashes, fees, values, amounts_matrix
-
-
-def process_loop_results(operator,
-                         loop_results=None,
-                         op_log_matrix=None,
-                         op_summary_matrix=None,
-                         op_list_bt_indices=None,
-                         trade_log=False,
-                         bt_price_priority_ohlc: str = 'OHLC'):
-    """ 接受apply_loop函数传回的计算结果，生成DataFrame型交易模拟结果数据，保存交易记录，并返回结果供下一步处理
-
-    Parameters
-    ----------
-    operator: Operator
-        交易操作对象
-    loop_results: tuple, optional
-        apply_loop函数返回的计算结果
-    op_log_matrix: np.ndarray, optional
-        交易记录矩阵，记录了模拟交易过程中每一笔交易的详细信息
-    op_summary_matrix: np.ndarray, optional
-        交易汇总矩阵，记录了模拟交易过程中每一笔交易的汇总信息
-    op_list_bt_indices: list, optional
-        交易记录矩阵中的交易日期索引
-    trade_log: bool, optional, default False
-        是否保存交易记录，默认为False
-    bt_price_priority_ohlc: str, optional, default 'OHLC'
-        交易记录中的价格优先级，可选'OHLC'、'HLOC'、'LOCH'、'LCOH'、'COHL'、'COLH'
-
-    Returns
-    -------
-    value_history: pd.DataFrame
-        交易模拟结果数据
-    """
-    from qteasy import logger_core
-    amounts_matrix, cashes, fees, values = loop_results
-    shares = operator.op_list_shares
-    complete_loop_dates = operator.op_list_hdates
-    looped_dates = [complete_loop_dates[item] for item in op_list_bt_indices]
-
-    price_types_in_priority = operator.get_bt_price_types_in_priority(priority=bt_price_priority_ohlc)
-
-    # 将向量化计算结果转化回DataFrame格式
-    value_history = pd.DataFrame(amounts_matrix, index=looped_dates,
-                                 columns=shares)
-    # 填充标量计算结果
-    value_history['cash'] = cashes
-    value_history['fee'] = fees
-    value_history['value'] = values
-
-    # 生成trade_log，index为MultiIndex，因为每天的交易可能有多种价格
-    if trade_log:
-        # create complete trading log
-        logger_core.info(f'generating complete trading log ...')
-        op_log_index = pd.MultiIndex.from_product(
-                [looped_dates,
-                 price_types_in_priority,
-                 ['0, trade signal',
-                  '1, price',
-                  '2, traded amounts',
-                  '3, cash changed',
-                  '4, trade cost',
-                  '5, own amounts',
-                  '6, available amounts',
-                  '7, summary']],
-                names=('date', 'trade_on', 'item')
-        )
-        op_sum_index = pd.MultiIndex.from_product(
-                [looped_dates,
-                 price_types_in_priority,
-                 ['7, summary']],
-                names=('date', 'trade_on', 'item')
-        )
-        op_log_columns = [str(s) for s in shares]
-        op_log_df = pd.DataFrame(op_log_matrix, index=op_log_index, columns=op_log_columns)
-        op_summary_df = pd.DataFrame(op_summary_matrix,
-                                     index=['add. invest', 'own cash', 'available cash', 'value'],
-                                     columns=op_sum_index).T
-        log_file_path_name = os.path.join(qteasy.QT_TRADE_LOG_PATH, 'trade_log.csv')
-        op_summary_df.join(op_log_df, how='right', sort=False).to_csv(log_file_path_name)
-
-        # 生成 trade log 摘要表 (a more concise and human-readable format of trading log
-        # create share trading logs:
-        logger_core.info(f'generating abstract trading log ...')
-        share_logs = []
-        for share in op_log_columns:
-            share_df = op_log_df[share].unstack()
-            share_df = share_df[share_df['2, traded amounts'] != 0]
-            share_df['code'] = share
-            try:
-                share_name = get_basic_info(share, printout=False)['name']
-            except Exception as e:
-                share_name = 'unknown'
-            share_df['name'] = share_name
-            share_logs.append(share_df)
-
-        re_columns = ['code',
-                      'name',
-                      '0, trade signal',
-                      '1, price',
-                      '2, traded amounts',
-                      '3, cash changed',
-                      '4, trade cost',
-                      '5, own amounts',
-                      '6, available amounts',
-                      '7, summary']
-        op_log_shares_abs = pd.concat(share_logs).reindex(columns=re_columns)
-        record_file_path_name = os.path.join(qteasy.QT_TRADE_LOG_PATH, 'trade_records.csv')
-        # TODO: 可以增加一个config属性来控制交易摘要表的生成规则：
-        #  如果how == 'left' 保留无交易日期的记录
-        #  如果how == 'right', 不显示无交易日期的记录
-        op_summary_df.join(op_log_shares_abs, how='right', sort=True).to_csv(record_file_path_name)
-
-    return value_history
-
 
 def filter_stocks(date: str = 'today', **kwargs) -> pd.DataFrame:
     """根据输入的参数筛选股票，并返回一个包含股票代码和相关信息的DataFrame
@@ -2997,26 +2036,9 @@ def run(operator, **kwargs):
     # 赋值给参考数据和运行模式
     benchmark_data_type = config['benchmark_asset']
     run_mode = config['mode']
-    """
-    用于交易信号生成、数据回测、策略优化、结果测试以及结果评价参考的历史数据:
-    hist_op:            信号生成数据，根据策略的运行方式，信号生成数据可能包含量价数据、基本面数据、指数、财务指标等等
-    back_trade_prices:          回测价格数据，用于在backtest模式下对生成的信号进行测试，仅包含买卖价格数据（定价回测）
-    hist_opti:          策略优化数据，数据的类型与信号生成数据一样，但是取自专门的独立区间用于策略的参数优化
-    hist_opti_loop:     优化回测价格，用于在optimization模式下在策略优化区间上回测交易结果，目前这组数据并未提前生成，
-                        而是在optimize的时候生成，实际上应该提前生成
-    hist_test:          策略检验数据，数据的类型与信号生成数据一样，同样取自专门的独立区间用于策略参数的性能检测
-    hist_test_loop:     检验回测价格，用于在optimization模式下在策略检验区间上回测交易结果
-    hist_benchmark:     评价参考价格，用于评价回测结果，大部分用于评价回测结果的alpha表现（取出无风险回报之后的表现）
-                        相关的指标都需要用到参考价格;
-    invest_cash_plan:
-    opti_cash_plan:
-    test_cssh_plan:
-    """
 
     if run_mode == 0 or run_mode == 'live':
-        '''进入实时信号生成模式: 
-        '''
-        # 启动交易shell
+        # 进入实时信号生成模式:
         from qteasy.trader import start_trader
         from qteasy import QT_DATA_SOURCE
         start_trader(
@@ -3069,6 +2091,7 @@ def run(operator, **kwargs):
 
     elif run_mode == 2 or run_mode == 'optimization':
         # 进入优化模式，使用真实历史数据或模拟历史数据反复测试策略，寻找并测试最佳参数
+
         # 判断operator对象的策略中是否有可优化的参数，即优化标记opt_tag设置为1，且参数数量不为0
         assert operator.opt_space_par[0] != [], \
             f'ConfigError, none of the strategy parameters is adjustable, set opt_tag to be 1 or 2 to ' \
@@ -3137,6 +2160,7 @@ def run(operator, **kwargs):
             if config['visual']:
                 _plot_test_result(test_eval_res=test_eval_res, opti_eval_res=opti_eval_res, config=config)
 
+        # TODO: Montecarlo模拟测试是否有必要存在？ v1.1版本后是否应该删除？
         elif config['test_type'] == 'montecarlo':
             for i in range(config['test_cycle_count']):
                 # 临时生成用于测试的模拟数据，将模拟数据传送到operator中，使用operator中的新历史数据
@@ -3169,717 +2193,3 @@ def run(operator, **kwargs):
                     pass
 
         return optimal_pars
-
-
-def _evaluate_all_parameters(par_generator,
-                             total,
-                             op: Operator,
-                             trade_price_list: HistoryPanel,
-                             benchmark_history_data,
-                             benchmark_history_data_type,
-                             config,
-                             stage='optimize') -> ResultPool:
-    """ 接受一个策略参数生成器对象，批量生成策略参数，反复调用_evaluate_one_parameter()函数，使用所有生成的策略参数
-        生成历史区间上的交易策略和回测结果，将得到的回测结果全部放入一个结果池对象，并根据策略筛选方法筛选出符合要求的回测
-        结果，并返回筛选后的结果。
-
-        根据config中的配置参数，这里可以选择进行并行计算以充分利用多核处理器的全部算力以缩短运行时间。
-
-    Parameters
-    ----------
-    par_generator: Iterables
-        一个迭代器对象，生成所有需要迭代测试的策略参数
-    op: Operator
-        一个operator对象，包含多个投资策略，用于根据交易策略以及策略的配置参数生成交易信号
-    trade_price_list: pd.DataFrame
-        用于进行回测的历史数据，该数据历史区间与前面的数据相同，但是仅包含回测所需要的价格信息，通常为收盘价
-        （假设交易价格为收盘价）
-    benchmark_history_data: pd.DataFrame
-        用于回测结果评价的参考历史数据，历史区间与回测历史数据相同，但是通常是能代表整个市场整体波动的金融资
-        产的价格，例如沪深300指数的价格。
-    benchmark_history_data_type: str
-        用于回测结果评价的参考历史数据种类，通常为收盘价close
-    config: Config
-        参数配置对象，用于保存相关配置，在所有的参数配置中，其作用的有下面N种：
-        1, config.opti_output_count:
-            优化结果数量
-        2, config.parallel:
-            并行计算选项，True时进行多进程并行计算，False时进行单进程计算
-    stage: str
-        该参数直接传递至_evaluate_one_parameter()函数中，其含义和作用参见其docstring
-
-    Returns
-    -------
-        pool，一个Pool对象，包含经过筛选后的所有策略参数以及它们的性能表现
-
-    """
-    pool = ResultPool(config.opti_output_count)  # 用于存储中间结果或最终结果的参数池对象
-    i = 0
-    best_so_far = 0
-    opti_target = config.optimize_target
-    # 启用多进程计算方式利用所有的CPU核心计算
-    if config.parallel:
-        # 启用并行计算
-        with ProcessPoolExecutor() as proc_pool:
-            futures = {proc_pool.submit(_evaluate_one_parameter,
-                                        par,
-                                        op,
-                                        trade_price_list,
-                                        benchmark_history_data,
-                                        benchmark_history_data_type,
-                                        config,
-                                        stage): par for par in
-                       par_generator}
-        for f in as_completed(futures):
-            eval_dict = f.result()
-            target_value = eval_dict[opti_target]
-            pool.in_pool(item=futures[f], perf=target_value, extra=eval_dict)
-            i += 1
-            if target_value > best_so_far:
-                best_so_far = target_value
-            if i % 10 == 0:
-                progress_bar(i, total, comments=f'best performance: {best_so_far:.3f}')
-    # 禁用多进程计算方式，使用单进程计算
-    else:
-        for par in par_generator:
-            perf = _evaluate_one_parameter(par=par,
-                                           op=op,
-                                           trade_price_list=trade_price_list,
-                                           benchmark_history_data=benchmark_history_data,
-                                           benchmark_history_data_type=benchmark_history_data_type,
-                                           config=config,
-                                           stage=stage)
-            target_value = perf[opti_target]
-            pool.in_pool(item=par, perf=target_value, extra=perf)
-            i += 1
-            if target_value > best_so_far:
-                best_so_far = target_value
-            if i % 10 == 0:
-                progress_bar(i, total, comments=f'best performance: {best_so_far:.3f}')
-    # 将当前参数以及评价结果成对压入参数池中，并返回所有成对参数和评价结果
-    progress_bar(i, i)
-
-    return pool
-
-
-def _evaluate_one_parameter(par,
-                            op: Operator,
-                            trade_price_list: HistoryPanel,
-                            benchmark_history_data,
-                            benchmark_history_data_type,
-                            config,
-                            stage='optimize') -> dict:
-    """ 基于op中的交易策略，在给定策略参数par的条件下，计算交易策略在一段历史数据上的交易信号，并对交易信号的交易
-        结果进行回测，对回测结果数据进行评价，并给出评价结果。
-
-    本函数是一个方便的包裹函数，包裹了交易信号生成、交易信号回测以及回测结果评价结果的打包过程，同时，根据QT基
-    本配置的不同，可以在交易信号回测的过程中进行多重回测，即将一段历史区间分成几个子区间，在每一个子区间上分别
-    回测后返回多次回测的综合结果。
-
-    Parameters
-    ----------
-    par: tuple, list, dict
-        输入的策略参数组合，这些参数必须与operator运行器对象中的交易策略相匹配，且符合op对象中每个交易策
-        略的优化标记设置，关于交易策略的优化标记如何影响参数导入，参见qt.operator.set_opt_par()的
-        docstring
-    op: qt.Operator
-        一个operator对象，包含多个投资策略，用于根据交易策略以及策略的配置参数生成交易信号
-    trade_price_list: HistoryPanel
-        用于模拟交易回测的历史价格，历史区间覆盖整个模拟交易期间，包含回测所需要的价格信息，可以为收盘价
-        和/或其他回测所需要的历史价格
-    benchmark_history_data: pd.DataFrame
-        用于回测结果评价的参考历史数据，历史区间与回测历史数据相同，但是通常是能代表整个市场整体波动的金融资
-        产的价格，例如沪深300指数的价格。
-    benchmark_history_data_type: str
-        用于回测结果评价的参考历史数据种类，通常为收盘价close，但也可以是其他价格，例如开盘价open
-    config: Config
-        参数配置对象，用于保存相关配置，在所有的参数配置中，其作用的有下面N种：
-        1, config.opti_type/test_type:
-            优化或测试模式，决定如何利用回测区间
-            single:     在整个回测区间上进行一次回测
-            multiple:   将回测区间分割为多个子区间并分别回测
-            montecarlo: 根据回测区间的数据生成模拟数据进行回测（仅在test模式下）
-        2, config.optimize_target/test_indicators:
-            优化目标函数（优化模式下）或评价指标（测试模式下）
-            在优化模式下，使用特定的优化目标函数来确定表现最好的策略参数
-            在测试模式下，对策略的回测结果进行多重评价并输出评价结果
-        3, config.opti_cash_amounts/test_cash_amounts:
-            优化/测试投资金额
-            在多区间回测情况下，投资金额会被调整，初始投资日期会等于每一个回测子区间的第一天
-        4, config.opti_sub_periods/test_sub_periods:
-            优化/测试区间数量
-            在多区间回测情况下，在整个回测区间中间隔均匀地取出多个区间，在每个区间上分别回测
-            每个区间的长度相同，但是起止点不同。每个起点之间的间隔与子区间的长度和数量同时相关，
-            确保每个区间的起点是均匀分布的，同时所有的子区间正好覆盖整个回测区间。
-        5, config.opti_sub_prd_length/test_sub_prd_length:
-            优化/测试子区间长度
-            该数值是一个相对长度，取值在0～1之间，代表每个子区间的长度相对于整个区间的比例，
-            例如，0.5代表每个子区间的长度是整个区间的一半
-    stage: str, optional, Default: 'optimize'
-        运行标记，代表不同的运行阶段控制运行过程的不同处理方式，包含三种不同的选项
-        1, 'loop':      运行模式为回测模式，在这种模式下：
-                        使用投资区间回测投资计划
-                        使用config.trade_log来确定是否打印回测结果
-        2, 'optimize':  运行模式为优化模式，在这种模式下：
-                        使用优化区间回测投资计划
-                        回测区间利用方式使用opti_type的设置值
-                        回测区间分段数量和间隔使用opti_sub_periods
-        3, 'test-o':    运行模式为测试模式-opti区间，以便在opti区间上进行一次与test区间完全相同的测试以比较结果
-                        使用优化区间回测投资计划
-                        回测区间利用方式使用test_type的设置值
-                        回测区间分段数量和间隔使用test_sub_periods
-        4, 'test-t':    运行模式为测试模式-test区间
-                        使用测试区间回测投资计划
-                        回测区间利用方式使用test_type的设置值
-                        回测区间分段数量和间隔使用test_sub_periods
-    Returns
-    -------
-    dict:
-    一个dict对象，存储该策略在使用par作为参数时的性能表现评分以及一些其他运行信息，允许对性能
-    表现进行多重指标评价，dict的指标类型为dict的键，评价结果为结果分值，dict不能为空，至少包含以下值：
-        'complete_value': 完整的回测结果清单，无结果时为None
-        'op_run_time':    交易清单生成耗时
-        'loop_run_time':  回测耗时
-        'final_value':    回测结果终值（默认评价指标）
-    除了上述必须存在的项目以外，返回的res_dict还可以包含任意evaluation模块可以输出的评价值，例如：
-        {'final_value': 34567,
-         'sharp':       0.123}
-    如果当前的策略不能生成有效的交易操作清单时，直接返回默认结果，其终值为负无穷大：
-        {'complete_values':  None,
-         'op_run_time':     0.0354675,
-         'loop_run_time':   None,
-         'final_value':     np.NINF}
-
-    """
-
-    res_dict = {'par':             None,
-                'complete_values': None,
-                'op_run_time':     0,
-                'loop_run_time':   0,
-                'final_value':     None}
-
-    assert stage in ['loop', 'optimize', 'test-o', 'test-t']
-    if par is not None:  # 如果给出了策略参数，则更新策略参数，否则沿用原有的策略参数
-        op.set_opt_par(par)
-        res_dict['par'] = par
-    # 生成交易清单并进行模拟交易生成交易记录
-    st = time.time()
-    op_list = None
-    if op.op_type == 'batch':
-        op_list = op.create_signal()
-    et = time.time()
-    op_run_time = et - st
-    res_dict['op_run_time'] = op_run_time
-    riskfree_ir = config.riskfree_ir
-    log_backtest = False
-    period_length = 0
-    period_count = 0
-    trade_dates = np.array(trade_price_list.hdates)
-    if op.op_type == 'batch' and op_list is None:  # 如果策略无法产生有意义的操作清单，则直接返回基本信息
-        res_dict['final_value'] = np.NINF
-        res_dict['complete_values'] = pd.DataFrame()
-        return res_dict
-    # 根据stage的值选择使用投资金额种类以及运行类型（单区间运行或多区间运行）及区间参数及回测参数
-    if stage == 'loop':
-        invest_cash_amounts = config.invest_cash_amounts
-        invest_cash_dates = pd.to_datetime(config.invest_start) if \
-            config.invest_cash_dates is None \
-            else pd.to_datetime(config.invest_cash_dates)
-        period_util_type = 'single'
-        indicators = 'years,fv,return,mdd,v,ref,alpha,beta,sharp,info'
-        log_backtest = config.trade_log  # 回测参数trade_log只有在回测模式下才有用
-    elif stage == 'optimize':
-        invest_cash_amounts = config.opti_cash_amounts[0]
-        # TODO: only works when config.opti_cash_dates is a string, if it is a list, it will not work
-        invest_cash_dates = pd.to_datetime(config.opti_start) if \
-            config.opti_cash_dates is None \
-            else pd.to_datetime(config.opti_cash_dates)
-        period_util_type = config.opti_type
-        period_count = config.opti_sub_periods
-        period_length = config.opti_sub_prd_length
-        indicators = config.optimize_target
-    elif stage == 'test-o':  # 在优化过程中，在优化区间上测试参数的性能
-        invest_cash_amounts = config.test_cash_amounts[0]
-        # TODO: only works when config.opti_cash_dates is a string, if it is a list, it will not work
-        invest_cash_dates = pd.to_datetime(config.opti_start) if \
-            config.opti_cash_dates is None \
-            else pd.to_datetime(config.opti_cash_dates)
-        period_util_type = config.test_type
-        period_count = config.test_sub_periods
-        period_length = config.test_sub_prd_length
-        indicators = config.test_indicators
-    else:  # stage == 'test-t':  # 在优化结束后，在测试区间上测试找到的最优参数的性能
-        invest_cash_amounts = config.test_cash_amounts[0]
-        # TODO: only works when config.opti_cash_dates is a string, if it is a list, it will not work
-        invest_cash_dates = pd.to_datetime(config.test_start) if \
-            config.test_cash_dates is None \
-            else pd.to_datetime(config.test_cash_dates)
-        period_util_type = config.test_type
-        period_count = config.test_sub_periods
-        period_length = config.test_sub_prd_length
-        indicators = config.test_indicators
-    # create list of start and end dates
-    # in this case, user-defined invest_cash_dates will be disabled, each start dates will be
-    # used as the investment date for each sub-periods
-    invest_cash_dates = next_market_trade_day(invest_cash_dates)
-    start_dates = []
-    end_dates = []
-    if period_util_type == 'single' or period_util_type == 'montecarlo':
-        start_dates.append(trade_dates[np.searchsorted(trade_dates, invest_cash_dates)])
-        end_dates.append(trade_dates[-1])
-    elif period_util_type == 'multiple':
-        # 多重测试模式，将一个完整的历史区间切割成多个区间，多次测试
-        first_history_date = invest_cash_dates
-        last_history_date = trade_dates[-1]
-        history_range = last_history_date - first_history_date
-        sub_hist_range = history_range * period_length
-        sub_hist_interval = (1 - period_length) * history_range / period_count
-        for i in range(period_count):
-            # 计算每个测试区间的起止点，抖动起止点日期，确保起止点在交易日期列表中
-            start_date = first_history_date + i * sub_hist_interval
-            start_date = trade_dates[np.searchsorted(trade_dates, start_date)]
-            start_dates.append(start_date)
-            end_date = start_date + sub_hist_range
-            end_date = trade_dates[np.searchsorted(trade_dates, end_date)]
-            end_dates.append(end_date)
-    else:
-        raise KeyError(f'Invalid optimization type: {config.opti_type}')
-    # loop over all pairs of start and end dates, get the results separately and output average
-    perf_list = []
-    price_priority_list = op.get_bt_price_type_id_in_priority(priority=config.price_priority_OHLC)
-    trade_cost = set_cost(
-            buy_fix=config.cost_fixed_buy,
-            sell_fix=config.cost_fixed_sell,
-            buy_rate=config.cost_rate_buy,
-            sell_rate=config.cost_rate_sell,
-            buy_min=config.cost_min_buy,
-            sell_min=config.cost_min_sell,
-            slipage=config.cost_slippage
-    )
-    st = time.time()
-    for start, end in zip(start_dates, end_dates):
-        start_idx = op.get_hdate_idx(start)
-        end_idx = op.get_hdate_idx(end)
-        trade_price_list_seg = trade_price_list.segment(start, end)
-        if stage != 'loop':
-            invest_cash_dates = trade_price_list_seg.hdates[0]
-        cash_plan = CashPlan(
-                invest_cash_dates.strftime('%Y%m%d'),
-                invest_cash_amounts,
-                riskfree_ir
-        )
-        # TODO: 将op_list_bt_indices的计算放到这一层函数中，以便下面几个函数可以共享
-        loop_results, op_log_matrix, op_summary_matrix, op_list_bt_indices = apply_loop(
-                operator=op,
-                trade_price_list=trade_price_list_seg,
-                start_idx=start_idx,
-                end_idx=end_idx,
-                cash_plan=cash_plan,
-                cost_rate=trade_cost,
-                moq_buy=config.trade_batch_size,
-                moq_sell=config.sell_batch_size,
-                inflation_rate=config.riskfree_ir,
-                pt_signal_timing=config.PT_signal_timing,
-                pt_buy_threshold=config.PT_buy_threshold,
-                pt_sell_threshold=config.PT_sell_threshold,
-                cash_delivery_period=config.cash_delivery_period,
-                stock_delivery_period=config.stock_delivery_period,
-                allow_sell_short=config.allow_sell_short,
-                long_pos_limit=config.long_position_limit,
-                short_pos_limit=config.short_position_limit,
-                max_cash_usage=config.maximize_cash_usage,
-                trade_log=log_backtest,
-                price_priority_list=price_priority_list
-        )
-        looped_val = process_loop_results(
-                operator=op,
-                loop_results=loop_results,
-                op_log_matrix=op_log_matrix,
-                op_summary_matrix=op_summary_matrix,
-                op_list_bt_indices=op_list_bt_indices,
-                trade_log=log_backtest,
-                bt_price_priority_ohlc='OHLC'
-        )
-        # TODO: 将_get_complete_hist() 与 process_loop_results()合并
-        complete_values = _get_complete_hist(
-                looped_value=looped_val,
-                h_list=trade_price_list,
-                benchmark_list=benchmark_history_data,
-                with_price=False
-        )
-        perf = evaluate(
-                looped_values=complete_values,
-                hist_benchmark=benchmark_history_data,
-                benchmark_data=benchmark_history_data_type,
-                cash_plan=cash_plan,
-                indicators=indicators
-        )
-        perf_list.append(perf)
-    perf = performance_statistics(perf_list)
-    et = time.time()
-    loop_run_time = et - st
-    res_dict.update(perf)
-    res_dict['loop_run_time'] = loop_run_time
-    return res_dict
-
-
-# TODO: 这个函数有潜在大量运行的可能，需要使用Numba加速
-def _create_mock_data(history_data: HistoryPanel) -> HistoryPanel:
-    """ 根据输入的历史数据的统计特征，随机生成多组具备同样统计特征的随机序列，用于进行策略收益的蒙特卡洛模拟
-
-    目前仅支持OHLC数据以及VOLUME数据的随机生成，其余种类的数据需要继续研究
-    为了确保生成的数据留有足够的前置数据窗口，生成的伪数据包含两段，第一段长度与最大前置窗口长度相同，这一段
-    为真实历史数据，第二段才是随机生成的模拟数据
-    同时，生成的数据仍然满足OHLC的关系，同时所有的数据在统计上与参考数据是一致的，也就是说，随机生成的数据
-    不仅仅满足K线图的形态要求，其各个参数的均值、标准差与参考数据一致。
-
-    Parameters
-    ----------
-    history_data: HistoryPanel
-        模拟数据的参考源
-
-    Returns
-    -------
-        HistoryPanel
-    """
-
-    assert isinstance(history_data, HistoryPanel)
-    data_types = history_data.htypes
-    # TODO: volume数据的生成还需要继续研究
-    assert any(data_type in ['close', 'open', 'high', 'low', 'volume'] for data_type in data_types), \
-        f'the data type {data_types} does not fit'
-    has_volume = any(data_type in ['volume'] for data_type in data_types)
-    # 按照细粒度方法同时生成OHLC数据
-    # 针对每一个share生成OHLC数据
-    # 先考虑生成正确的信息，以后再考虑优化
-    dfs_for_share = []
-    for share in history_data.shares:
-        share_df = history_data.slice_to_dataframe(share=share)
-        share_df['close_chg'] = share_df.close / share_df.close.shift(1)
-        mean = share_df.close_chg.mean()
-        std = share_df.close_chg.std()
-        mock_col = np.random.randn(len(history_data.hdates) * 5) * std * 5 + mean
-        mock_col = 1 + 0.09 * (mock_col - 1)
-        mock_col[0] = share_df.close.iloc[0]
-        mock_col = np.cumprod(mock_col)
-        mock = mock_col.reshape(len(history_data.hdates), 5)
-        mock_df = pd.DataFrame(index=history_data.hdates)
-        mock_df['open'] = mock[:, 0]
-        mock_df['high'] = np.max(mock, axis=1)
-        mock_df['low'] = np.min(mock, axis=1)
-        mock_df['close'] = mock[:, 4]
-        if has_volume:
-            mock_df['volume'] = share_df.volume
-        dfs_for_share.append(mock_df.copy())
-
-    # 生成一个HistoryPanel对象，每一层一个个股
-    mock_data = stack_dataframes(dfs_for_share,
-                                 dataframe_as='shares',
-                                 shares=history_data.shares)
-    return mock_data
-
-
-def _search_grid(hist, benchmark, benchmark_type, op, config):
-    """ 最优参数搜索算法1: 网格搜索法
-
-    在整个参数空间中建立一张间距固定的"网格"，搜索网格的所有交点所在的空间点，
-    根据该点的参数生成操作信号、回测后寻找表现最佳的一组或多组参数
-    与该算法相关的设置选项有：
-    grid_size:  网格大小，float/int/list/tuple 当参数为数字时，生成空间所有方向
-                上都均匀分布的网格；当参数为list或tuple时，可以在空间的不同方向
-                上生成不同间隔大小的网格。list或tuple的维度须与空间的维度一致
-
-    Parameters
-    ----------
-    hist: HistoryPanel
-        历史数据，优化器的整个优化过程在历史数据上完成
-    op: qt.Operator
-        交易信号生成器对象
-    config: qt.Config
-        用于存储优化参数配置变量
-
-    Returns
-    -------
-    tuple: (pool.items, pool.perfs)
-        pool.items 作为结果输出的参数组
-        pool.perfs 输出的参数组的评价分数
-    """
-    s_range, s_type = op.opt_space_par
-    space = Space(s_range, s_type)  # 生成参数空间
-
-    # 使用extract从参数空间中提取所有的点，并打包为iterator对象进行循环
-    par_generator, total = space.extract(config.opti_grid_size)
-    history_list = hist.fillna(0)
-    st = time.time()
-    pool = _evaluate_all_parameters(par_generator=par_generator,
-                                    total=total,
-                                    op=op,
-                                    trade_price_list=history_list,
-                                    benchmark_history_data=benchmark,
-                                    benchmark_history_data_type=benchmark_type,
-                                    config=config,
-                                    stage='optimize')
-    pool.cut(config.maximize_target)
-    et = time.time()
-    print(f'\nOptimization completed, total time consumption: {sec_to_duration(et - st)}')
-    return pool.items, pool.perfs
-
-
-def _search_montecarlo(hist, benchmark, benchmark_type, op, config):
-    """ 最优参数搜索算法2: 蒙特卡洛法
-
-        从待搜索空间中随机抽取大量的均匀分布的参数点并逐个测试，寻找评价函数值最优的多个参数组合
-        与该算法相关的设置选项有：
-            sample_size:采样点数量，int 由于采样点的分布是随机的，因此采样点越多，越有可能
-                        接近全局最优值
-
-    Parameters
-    ----------
-    hist: HistoryPanel
-        历史数据，优化器的整个优化过程在历史数据上完成
-    benchmark:
-        基准数据，用于计算基准收益率
-    benchmark_type:
-        基准数据类型，用于计算基准收益率
-    op: qt.Operator
-        交易信号生成器对象
-    config: qt.Config
-        用于存储交易相关参数的配置变量
-
-    Returns
-    -------
-    tuple: (pool.items, pool.perfs)
-        pool.items 作为结果输出的参数组
-        pool.perfs 输出的参数组的评价分数
-    """
-    # s_range, s_type = a_to_sell.opt_space_par
-    space = Space(*op.opt_space_par)  # 生成参数空间
-    # 使用随机方法从参数空间中取出point_count个点，并打包为iterator对象，后面的操作与网格法一致
-    par_generator, total = space.extract(config.opti_sample_count, how='rand')
-    history_list = hist.fillna(0)
-    st = time.time()
-    pool = _evaluate_all_parameters(par_generator=par_generator,
-                                    total=total,
-                                    op=op,
-                                    trade_price_list=history_list,
-                                    benchmark_history_data=benchmark,
-                                    benchmark_history_data_type=benchmark_type,
-                                    config=config,
-                                    stage='optimize')
-    pool.cut(config.maximize_target)
-    et = time.time()
-    print(f'\nOptimization completed, total time consumption: {sec_to_duration(et - st, short_form=True)}')
-    return pool.items, pool.perfs
-
-
-def _search_incremental(hist, benchmark, benchmark_type, op, config):
-    """ 最优参数搜索算法3: 增量递进搜索法
-
-    该算法是蒙特卡洛算法的一种改进。整个算法运行多轮蒙特卡洛算法，但是每一轮搜索的空间大小都更小，
-    而且每一轮搜索都（大概率）更接近全局最优解。
-    该算法的第一轮搜索就是标准的蒙特卡洛算法，在整个参数空间中随机取出一定数量的参数组合，使用这
-    些参数分别进行信号回测。第一轮搜索结束后，在第一轮的全部结果中择优选出一定比例的最佳参数，以
-    这些最佳参数为中心点，构建一批子空间，这些子空间的总体积比起最初的参数空间小的多，但是大概率
-    容纳了最初参数空间的全局最优解。
-    接着，程序继续在新生成的子空间中取出同样多的参数组合，并同样选出最佳参数组合，以新的最优解为
-    中心创建下一轮的参数空间，其总体积再次缩小。
-    如上所诉反复运行程序，每一轮需要搜索的子空间的体积越来越小，找到全局最优的概率也越来越大，直到
-    参数空间的体积小于一个固定值，或者循环的次数超过最大次数，循环停止，输出当前轮的最佳参数组合。
-
-    与该算法相关的设置选项有：
-        r_sample_size:      采样点数量，int 每一轮搜索中采样点的数量
-        reduce_ratio:       择优比例，float, 大于零小于1的浮点数，次轮搜索参数空间大小与本轮
-                            空间大小的比例，同时也是参数组的择优比例，例如0。2代表每次搜索的
-                            参数中最佳的20%会被用于创建下一轮的子空间邻域，同时下一轮的子空间
-                            体积为本轮空间体积的20%
-        max_rounds:         最大轮数，int，循环次数达到该值时结束循环
-        min_volume:         最小体积，float，当参数空间的体积（Volume）小于该值时停止循环
-
-    Parameters
-    ----------
-    hist: HistoryPanel
-        历史数据，优化器的整个优化过程在历史数据上完成
-    op: qt.Operator
-        交易信号生成器对象
-    config: qt.Config
-        用于存储交易相关参数的配置变量
-
-    Returns
-    -------
-    tuple，包含两个变量
-        pool.items 作为结果输出的参数组
-        pool.perfs 输出的参数组的评价分数
-    """
-    sample_count = config.opti_r_sample_count
-    min_volume = config.opti_min_volume
-    max_rounds = config.opti_max_rounds
-    reduce_ratio = config.opti_reduce_ratio
-    parallel = config.parallel
-    s_range, s_type = op.opt_space_par
-    spaces = list()  # 子空间列表，用于存储中间结果邻域子空间，邻域子空间数量与pool中的元素个数相同
-    base_space = Space(s_range, s_type)
-    base_volume = base_space.volume
-    base_dimension = base_space.dim
-    # 每一轮参数寻优后需要保留的参数组的数量
-    reduced_sample_count = int(sample_count * reduce_ratio)
-    pool = ResultPool(reduced_sample_count)  # 用于存储中间结果或最终结果的参数池对象
-
-    spaces.append(base_space)  # 将整个空间作为第一个子空间对象存储起来
-    space_count_in_round = 1  # 本轮运行子空间的数量
-    current_round = 1  # 当前运行轮次
-    current_volume = base_space.volume  # 当前运行轮次子空间的总体积
-    history_list = hist.fillna(0)  # 准备历史数据
-    """
-    估算运行的总回合数量，由于每一轮运行的回合数都是大致固定的（随着空间大小取整会有波动）
-    因此总的运行回合数就等于轮数乘以每一轮的回合数。关键是计算轮数
-    由于轮数的多少取决于两个变量，一个是最大轮次数，另一个是下一轮产生的子空间总和体积是否
-    小于最小体积阈值，因此，推算过程如下：
-    设初始空间体积为Vi，最小空间体积为Vmin，每一轮的缩小率为rr，最大计算轮数为Rmax
-    且第k轮的空间体积为Vk，则有：
-                          Vk = Vi * rr ** k
-          停止条件1：      Vk = Vi * rr ** k < Vmin
-          停止条件2:      k >= Rmax
-       根据停止条件1：    rr ** k < Vmin / Vi
-                       k > log(Vmin / Vi) / log(rr)
-           因此，当：    k > min(Rmax, log(Vmin / Vi) / log(rr))
-    """
-    round_count = min(max_rounds, (math.log(min_volume / base_volume) / math.log(reduce_ratio)))
-    total_calc_rounds = int(round_count * sample_count)
-    i = 0
-    st = time.time()
-    # 从当前space开始搜索，当subspace的体积小于min_volume或循环次数达到max_rounds时停止循环
-    while current_volume >= min_volume and current_round < max_rounds:
-        # 在每一轮循环中，spaces列表存储该轮所有的空间或子空间
-        while spaces:
-            space = spaces.pop()
-            # 逐个弹出子空间列表中的子空间，随机选择参数，生成参数生成器generator
-            # 生成的所有参数及评价结果压入pool结果池，每一轮所有空间遍历完成后再排序择优
-            par_generator, total = space.extract(sample_count // space_count_in_round, how='rand')
-            # TODO: progress bar does not work properly, try to find a way to get progress bar working
-            pool = pool + _evaluate_all_parameters(par_generator=par_generator,
-                                                   total=total,
-                                                   op=op,
-                                                   trade_price_list=history_list,
-                                                   benchmark_history_data=benchmark,
-                                                   benchmark_history_data_type=benchmark_type,
-                                                   config=config,
-                                                   stage='optimize')
-        # 本轮所有结果都进入结果池，根据择优方向选择最优结果保留，剪除其余结果
-        pool.cut(config.maximize_target)
-        """
-        为了生成新的子空间，计算下一轮子空间的半径大小
-        为确保下一轮的子空间总体积与本轮子空间总体积的比值是reduce_ratio，需要根据空间的体积公式设置正确
-        的缩小比例。这个比例与空间的维数和子空间的数量有关
-        例如：
-        若 reduce_ratio(rr)=0.5，设初始空间体积为Vi,边长为Si，第k轮空间体积为Vk，子空间数量为m，
-              每个子空间的体积为V，Size为S，空间的维数为d,则有：
-              Si ** d * (rr ** k) = Vi * (rr ** k) = Vk =  V * m = S ** d * m
-              于是：
-              S ** d * m = Si ** d * (rr ** k)
-              (S/Si) ** d = (rr ** k) / m
-              S/Si = ((rr ** k) / m) ** (1/d)
-        根据上述结果，第k轮的子空间直径S可以由原始空间的半径Si得到：
-              S = Si * ((rr ** k) / m) ** (1/d)
-              distance = S / 2
-        """
-        size_reduce_ratio = ((reduce_ratio ** current_round) / reduced_sample_count) ** (1 / base_dimension)
-        reduced_size = tuple(np.array(base_space.size) * size_reduce_ratio / 2)
-        # 完成一轮搜索后，检查pool中留存的所有点，并生成由所有点的邻域组成的子空间集合
-        current_volume = 0
-        for point in pool.items:
-            subspace = base_space.from_point(point=point, distance=reduced_size)
-            spaces.append(subspace)
-            current_volume += subspace.volume
-        current_round += 1
-        space_count_in_round = len(spaces)
-        progress_bar(i, total_calc_rounds, f'start next round with {space_count_in_round} spaces')
-    et = time.time()
-    print(f'\nOptimization completed, total time consumption: {sec_to_duration(et - st)}')
-    return pool.items, pool.perfs
-
-
-def _search_ga(hist, benchmark, benchmark_type, op, config):
-    """ 最优参数搜索算法4: 遗传算法
-    遗传算法适用于在超大的参数空间内搜索全局最优或近似全局最优解，而它的计算量又处于可接受的范围内
-
-    遗传算法借鉴了生物的遗传迭代过程，首先在参数空间中随机选取一定数量的参数点，将这批参数点称为
-    “种群”。随后在这一种群的基础上进行迭代计算。在每一次迭代（称为一次繁殖）前，根据种群中每个个体
-    的评价函数值，确定每个个体生存或死亡的几率，规律是若个体的评价函数值越接近最优值，则其生存的几率
-    越大，繁殖后代的几率也越大，反之则越小。确定生死及繁殖的几率后，根据生死几率选择一定数量的个体
-    让其死亡，而从剩下的（幸存）的个体中根据繁殖几率挑选几率最高的个体进行杂交并繁殖下一代个体，
-    同时在繁殖的过程中引入随机的基因变异生成新的个体。最终使种群的数量恢复到初始值。这样就完成
-    一次种群的迭代。重复上面过程数千乃至数万代直到种群中出现希望得到的最优或近似最优解为止
-
-    Parameters
-    ----------
-    hist: HistoryPanel
-        历史数据，优化器的整个优化过程在历史数据上完成
-    benchmark:
-
-    benchmark_type:
-
-    op: object，
-        交易信号生成器对象
-    config: ConfigDict
-
-    Returns
-    -------
-    tuple，包含两个变量
-        pool.items 作为结果输出的参数组
-        pool.perfs 输出的参数组的评价分数
-
-    """
-    raise NotImplementedError
-
-
-def _search_gradient(hist, benchmark, benchmark_type, op, config):
-    """ 最优参数搜索算法5：梯度下降法
-    在参数空间中寻找优化结果变优最快的方向，始终保持向最优方向前进（采用自适应步长）一直到结果不再改变或达到
-    最大步数为止，输出结果为最后N步的结果
-
-    Parameters
-    ----------
-    hist，object，历史数据，优化器的整个优化过程在历史数据上完成
-    benchmark:
-    benchmark_type:
-    op: object，交易信号生成器对象
-    config: object, 用于存储交易相关参数配置对象
-
-    Returns
-    -------
-    """
-    raise NotImplementedError
-
-
-def _search_pso(hist, benchmark, benchmark_type, op, config):
-    """ Particle Swarm Optimization 粒子群优化算法，与梯度下降相似，从随机解出发，通过迭代寻找最优解
-
-    Parameters
-    ----------
-    hist，object，历史数据，优化器的整个优化过程在历史数据上完成
-    benchmark:
-    benchmark_type:
-    op: object，交易信号生成器对象
-    config: object, 用于存储交易相关参数配置对象
-
-    Returns
-    -------
-    """
-    raise NotImplementedError
-
-
-def _search_aco(hist, benchmark, benchmark_type, op, config):
-    """ Ant Colony Optimization 蚁群优化算法，
-
-    Parameters
-    ----------
-    hist，object，历史数据，优化器的整个优化过程在历史数据上完成
-    benchmark:
-    benchmark_type:
-    op，object，交易信号生成器对象
-    config: object, 用于存储交易相关参数配置对象
-
-    Returns
-    -------
-    """
-    raise NotImplementedError
